Move the FilterTrialsButton in a specific file
